requirements.txt
google-generative-ai
requests
beautifulsoup4
python-dotenv